requirements.txt
python==3.12.0
asynctest==0.13.0
pymongo==4.6.1
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-mock-resources==2.10.0
pytest-tornasync==0.6.0.post2
python-dotenv==1.0.0
python-telegram-bot==20.7
requests==2.31.0
requests-toolbelt==1.0.0
setuptools==69.0.3